requirements.txt
python-telegram-bot==20.3
feedparser==6.0.10
python-dotenv==1.0.0